[WIP] create /systemd, move stats' service template there, and create timer
I'm switching from daemon service to oneshot service scheculed by a timer due
to the high RAM usage when the daemon is asleep. (A Python interpreter seems to
consume at least ~9MB of RAM on my Ubuntu 16.04 x64 instance; the stats daemon
in particular consumes a whopping ~50MB, probably due to all the matplotlib
stuff -- maybe there were leaks or uncollected garbage, I didn't bother to dig
into that, because even 9MB during sleep sounds a bit too much, and I have
multiple such daemons).

A Makefile is added to automate the building and installation of the systemd
units.

This is a WIP; I've yet to migrate other components to this regime, and
documentation has not caught up yet.