stats: mark data.csv link as target="_blank" rather than download
Forcing the download of a file that is plain text in nature is a dick
move. I'm changing the server config too to serve data.csv as text/plain
rather than text/csv, so that the file is opened as a plain text tab in
browsers, and users can decide whether they want to download it, copy
from it, or whatever.