Always expand to realpath when obtaining script directory
For some reason I got an ImportError in mod_wsgi when loading
schedule/app.wsgi, due to the line

    from app import init

and I've no idea why. The only thing I changed recently is to change the
WSGIProcessGroup of each of my WSGIDaemonProcess to %{GLOBAL} instead of
one for each. Anyway, using realpath seems to have fixed it, or made it
unreproducible, at least for now...